+++ stustapay/core/schema/customer.py
@@ -1,6 +1,7 @@
 import datetime
 import uuid
 from typing import Optional
+from decimal import Decimal
 
 from pydantic import BaseModel
 
@@ -13,8 +14,8 @@ class Customer(Account):
     iban: Optional[str]
     account_name: Optional[str]
     email: Optional[str]
-    donation: Optional[float]
-    payout_amount: Optional[float]
+    donation: Optional[Decimal]
+    payout_amount: Optional[Decimal]
     payout_error: Optional[str]
     payout_run_id: Optional[int]
     payout_export: Optional[bool]
@@ -26,7 +27,7 @@ class OrderWithBon(Order):
 
 class CustomerCheckout(BaseModel):
     checkout_reference: uuid.UUID
-    amount: float
+    amount: Decimal
     currency: str
     merchant_code: str
     description: str
