+++ stustapay/core/schema/ticket.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from pydantic import BaseModel
 
 from stustapay.core.schema.product import ProductRestriction
@@ -5,11 +7,11 @@ from stustapay.core.schema.product import ProductRestriction
 
 class NewTicket(BaseModel):
     name: str
-    price: float
+    price: Decimal
     tax_rate_id: int
     restrictions: list[ProductRestriction]
     is_locked: bool
-    initial_top_up_amount: float
+    initial_top_up_amount: Decimal
 
 
 class Ticket(NewTicket):
@@ -17,5 +19,5 @@ class Ticket(NewTicket):
     id: int
 
     tax_name: str
-    tax_rate: float
-    total_price: float
+    tax_rate: Decimal
+    total_price: Decimal
