+++ stustapay/core/service/till/register.py
@@ -1,4 +1,5 @@
 from typing import Optional
+from decimal import Decimal
 
 import asyncpg
 
@@ -306,7 +307,7 @@ class TillRegisterService(DBService):
         node: Node,
         current_user: CurrentUser,
         cashier_tag_uid: int,
-        amount: float,
+        amount: Decimal,
     ):
         # TODO: TREE visibility
         row = await conn.fetchrow(
@@ -339,7 +340,7 @@ class TillRegisterService(DBService):
                 f"Insufficient balance on transport account. Current balance is {transport_account.balance}"
             )
 
-        bookings: dict[BookingIdentifier, float] = {
+        bookings: dict[BookingIdentifier, Decimal] = {
             BookingIdentifier(source_account_id=transport_account.id, target_account_id=cashier_account.id): amount
         }
 
@@ -363,7 +364,7 @@ class TillRegisterService(DBService):
         current_user: CurrentUser,
         current_terminal: CurrentTerminal,
         orga_tag_uid: int,
-        amount: float,
+        amount: Decimal,
     ):
         # TODO: TREE visibility
         transport_account = await get_transport_account_by_tag_uid(conn=conn, orga_tag_uid=orga_tag_uid)
